add utility function for verifying simulator output directory
use this in fvbaseproblem.hh. allows to encapsulate some includes
