+++ opm/models/discretization/common/fvbaseproblem.hh
@@ -44,9 +44,6 @@
 #include <limits>
 #include <string>
 
-#include <sys/stat.h>
-#include <unistd.h>
-
 namespace Opm::Properties {
 
 template <class TypeTag, class MyTypeTag>
@@ -208,23 +205,7 @@ public:
      */
     std::string outputDir() const
     {
-        std::string outputDir = Parameters::Get<Parameters::OutputDir>();
-
-        if (outputDir.empty())
-            outputDir = ".";
-
-        // TODO: replace this by std::filesystem once we require c++-2017
-        struct stat st;
-        if (::stat(outputDir.c_str(), &st) != 0)
-            throw std::runtime_error("Could not access output directory '"+outputDir+"':"
-                                     +strerror(errno));
-        if (!S_ISDIR(st.st_mode))
-            throw std::runtime_error("Path to output directory '"+outputDir+"' exists but is not a directory");
-
-        if (access(outputDir.c_str(), W_OK) != 0)
-            throw std::runtime_error("Output directory '"+outputDir+"' exists but is not writeable");
-
-        return outputDir;
+        return simulatorOutputDir();
     }
 
     /*!
